+++ src/main/java/com/example/demo/repositories/RequestRepository.java
@@ -10,11 +10,11 @@ import java.util.List;
 
 public interface RequestRepository extends CrudRepository<Request, Serializable> {
     @Query(value = "SELECT * FROM social_network.requests WHERE id = ?1", nativeQuery = true)
-    Request getRequestById(int id);
+    Request getRequestById(long id);
 
     @Query(value = "SELECT * FROM social_network.requests WHERE sender_id = ?1 AND receiver_id = ?2", nativeQuery = true)
-    Request existsByUniquePair(int sender, int receiver);
+    Request existsByUniquePair(long sender, long receiver);
 
     @Query(value = "SELECT * from social_network.requests where receiver_id = ?1", nativeQuery = true)
-    List<Request> getUserRequests(int userId);
+    List<Request> getUserRequests(long userId);
 }
